tf-keras/requirements.txt
awkward>=0.12.0
numpy>=1.14.0
pandas>=0.23.4
uproot-methods>=0.7.0
tensorflow-gpu>=2.0.0
